implemented dummy indexed storage and moved AbstruseIndex to storage package
