+++ aitools/proofs/knowledge_bases/redis.py
@@ -1,12 +1,11 @@
 import pickle
-from collections import OrderedDict
 from typing import Optional, Iterable, Set, Collection
 
 import redis
 
 from aitools.logic import Expression, Substitution, Variable
 from aitools.logic.utils import VariableSource, normalize_variables
-from aitools.proofs.index import make_key
+from aitools.storage.index import make_key
 from aitools.utils.abstruse_index import TrieIndex, AbstruseIndex
 from aitools.proofs.knowledge_bases.knowledge_base import KnowledgeBase
 from aitools.proofs.listeners import Listener
